fix: issues after implement code review recommendations
