+++ src/components/ui/button.tsx
@@ -9,6 +9,11 @@ const buttonVariants = cva(
   'inline-flex items-center justify-center whitespace-nowrap rounded-md text-base line-height-[28px] font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
   {
     variants: {
+      size: {
+        default: 'px-5 py-3',
+        sm: 'rounded-md px-4 py-2',
+        lg: 'rounded-md px-8',
+      },
       variant: {
         default: 'bg-primary text-white hover:bg-primary-300 hover:border-primary-300 active:bg-primary-400 active:border-primary-400',
         destructive:
@@ -21,12 +26,7 @@ const buttonVariants = cva(
         fullLink: 'block text-grey-500 w-full py-4 text-center mb-3 rounded-lg',
         link: 'text-primary underline-offset-4 hover:underline',
         pillPrimary: 'border-2 border-primary rounded-full text-primary font-bold hover:bg-primary-300 hover:border-primary-300 active:bg-primary-400  active:border-primary-400 hover:text-white',
-        icon: 'flex w-11 h-11 bg-primary-100 color-primary border-none rounded-full hover:bg-primary-300 active:bg-primary-400',
-      },
-      size: {
-        default: 'px-5 py-3',
-        sm: 'rounded-md px-4 py-2',
-        lg: 'rounded-md px-8',
+        icon: 'flex w-11 h-11 bg-primary-100 color-primary border-none rounded-full hover:bg-primary-300 active:bg-primary-400 p-0',
       },
     },
     defaultVariants: {
